+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Modal.set({ buttons: ... }) now also accepts new buttons as strings
     - Fix and improve buttons definition and reactivity
     - CANCEL and CLOSE buttons default to dismiss=true
+    - Submit the modal on Enter
 
 ### 1.8.0